*** empty log message ***
svn path=/trunk/boinc/; revision=1646

+++ lib/util.C
@@ -478,6 +478,6 @@ string timediff_format(long tdiff)
         return buf;
     }
 
-    sprintf(buf, "%d weeks, %d days, %d hours, %d minutes, and %d seconds", tdiff, days, hours, min, sex);
+    sprintf(buf, "%d weeks, %d days, %d hours, %d minutes, and %d seconds", (int)tdiff, days, hours, min, sex);
     return buf;
 }
